Complete Google Cloud integration
Uses the custom runtime environment, based off of a GCE OpenJDK Dockerfile.  Update Kotlin, remove Gradle Docker plugin.  Add Google Cloud endpoints.  Don't use old package name.
