+++ app.yaml
@@ -1,5 +1,6 @@
-runtime: java
+runtime: custom
 env: flex
 
+# Unsure if this is needed, from: https://cloud.google.com/appengine/docs/flexible/java/configuring-your-app-with-app-yaml
 runtime_config:
   jdk: openjdk8
