+++ readme.md
@@ -4,8 +4,6 @@
   * https://projects.spring.io/spring-boot/
 * OpenAPI (aka Swagger) integration via SpringFox
   * http://springfox.github.io/springfox/
-* Build docker image
-  * https://spring.io/guides/gs/spring-boot-docker/
 
 ## Run locally
 
@@ -17,7 +15,7 @@ Swagger: `http://localhost:8080/swagger-ui.html`
 
 ## Docker
 
-Build docker image: `./gradlew buildDocker`
+Build docker image: `docker build -t spring-booty:latest .`
 
 Run docker image: `docker run -p 8080:8080 spring-booty:latest`
 
@@ -33,11 +31,6 @@ Remove dangling images: `docker rmi $(docker images --filter "dangling=true" -q 
     * https://spring.io/guides/gs/rest-service/
 3. Add swagger integration
     * http://heidloff.net/article/usage-of-swagger-2-0-in-spring-boot-applications-to-document-apis/
-4. Add docker integration
-    * https://geowarin.github.io/spring-boot-app-in-docker-image.html
-5. Added Kotlin integration
-
-## Next
-
-* Jenkins configuration in project
-* Configuration service / Spring Cloud Consul
+4. Add Dockerfile
+5. Add Kotlin
+6. Add Google Cloud
